add: cloud natural language api 사용을 위한 config 설정

+++ src/main/java/LearnMate/dev/model/converter/DiaryConverter.java
@@ -28,7 +28,7 @@ public class DiaryConverter {
                 .build();
     }
 
-    public static DiaryAnalysisResponse toDiaryAnalysisResponse(Double score, String actionTip) {
+    public static DiaryAnalysisResponse toDiaryAnalysisResponse(Float score, String actionTip) {
         return DiaryAnalysisResponse.builder()
                 .emotionScore(score)
                 .actionTip(actionTip)
